backend/requirements.txt
google-generativeai>=0.3.0
django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
requests>=2.31.0
django-cors-headers>=4.3.0 